formatting and structural arrangement changes
changed to match the sections that was mention by the lecturer.

change the '#' , "##', and '###' so that the headings are correctly written

rearranged the prop.test code and the data loading code chunk to match the flow of the written text.

deleted code that loads libraries and put them at the beginning of the rmarkdown file.